refactor: resolve #326 - split FunctionEvaluator.test.ts from 811 to under 500 lines
Closes #326
